engine: Add Move data and basic makeMove function
This change introduces the abstraction of a move as a differential
change from some state of the board. The Move object keeps track
of the information needed to apply the move, as well as some data
usefull for further move evaluation.

A makeMove function is also implemented, allowing to change any
board state by the application of the given move. This function produces
a valid board state - the result of playing the supplied move.

Currently there's no method for undoing moves.