engine: Add a naive search function for determining best moves
This change introduces a Search module exposing a best move calculating
function called `search`. It assumes two players, white and black, are
trying to optimize the game score, being the max player and min player
respectively.
For now, the search is a simple evaluation of currently possible moves,
terminating at depth 1 and returning the best one according to the
current
player's optimization function.