interfaces: Don't send moves exposing the king to the client
For the purposes of the engine we only generate pseudo legal
moves as there is no need for the explicit check - the engine
will never want to lose the king.
However, in the exposed interface for the legal moves we should
filter out the moves that leave the king in check.